fix: correctly cache and route the svelte-kit routes
Svelte-kit routes have no '.html' matching files and do not appear
in the kit's provided list of files or build. This change adds a
variable where the routes can be listed to be added to the list for
caching and each is given a route with StaleWhileRevalidate
strategy, allowing the cache to be updated yet serving the cached
value until the new version is loaded.